Refactor project with mvc estrategy
