+++ .gitignore
@@ -36,3 +36,8 @@ build/
 
 ### Mac OS ###
 .DS_Store
+
+# Archivos de configuración con credenciales
+.env
+src/main/resources/config.properties
+config.properties
