+++ src/main/java/com/biblioteca/model/dao/ConexionBD.java
@@ -1,0 +1,37 @@
+package com.biblioteca.model.dao;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class ConexionBD {
+    private static final String URL = "jdbc:mysql://localhost:3306/biblioteca";
+    private static final String USER = "root";
+    private static final String PASSWORD = "pass";
+
+    private static Connection instance;
+
+    private ConexionBD() {}
+
+    public static Connection getConnection() throws SQLException {
+        if (instance == null || instance.isClosed()) {
+            try {
+                Class.forName("com.mysql.cj.jdbc.Driver");
+                instance = DriverManager.getConnection(URL, USER, PASSWORD);
+            } catch (ClassNotFoundException e) {
+                throw new SQLException("Driver MySQL no encontrado", e);
+            }
+        }
+        return instance;
+    }
+
+    public static void closeConnection() {
+        try {
+            if (instance != null && !instance.isClosed()) {
+                instance.close();
+            }
+        } catch (SQLException e) {
+            System.err.println("Error al cerrar la conexión: " + e.getMessage());
+        }
+    }
+}
